move alpha inside rl-regressor class
Found a way to create alpha inside the class that seems to work. Had to rename the function for creating knodes (create_stochastic_knodes_rl) and call that in the already existing create_stochastic_knodes in the HDDMrlRegressor. Also had to call to the new function to create the Intercept-knodes.